+++ src/redteam/util.ts
